修改记录2017.07.27   #{      @Time 14:00 修改上传图片使用方案   #}

+++ Sampling/app/src/main/java/com/xiangxun/sampling/ui/setting/SettingFragment.java
@@ -248,15 +248,17 @@ public class SettingFragment extends BaseFunctionList implements SettingFragment
             return;
         }
         if (file.isDirectory()) {
-            File[] childFile = file.listFiles();
-            if (childFile == null || childFile.length == 0) {
+            if (!file.getAbsolutePath().contains("video")) {
+                File[] childFile = file.listFiles();
+                if (childFile == null || childFile.length == 0) {
+                    file.delete();
+                    return;
+                }
+                for (File f : childFile) {
+                    RecursionDeleteFile(f);
+                }
                 file.delete();
-                return;
             }
-            for (File f : childFile) {
-                RecursionDeleteFile(f);
-            }
-            file.delete();
         }
     }
